+++ pkg/apis/rocketmq/v1alpha1/doc.go
@@ -17,5 +17,5 @@
 
 // Package v1alpha1 contains API Schema definitions for the rocketmq v1alpha1 API group
 // +k8s:deepcopy-gen=package,register
-// +groupName=rocketmq.operator.com
+// +groupName=rocketmq.apache.org
 package v1alpha1
