+++ pkg/apis/rocketmq/v1alpha1/register.go
@@ -19,7 +19,7 @@
 
 // Package v1alpha1 contains API Schema definitions for the rocketmq v1alpha1 API group
 // +k8s:deepcopy-gen=package,register
-// +groupName=rocketmq.operator.com
+// +groupName=rocketmq.apache.org
 package v1alpha1
 
 import (
@@ -29,7 +29,7 @@ import (
 
 var (
 	// SchemeGroupVersion is group version used to register these objects
-	SchemeGroupVersion = schema.GroupVersion{Group: "rocketmq.operator.com", Version: "v1alpha1"}
+	SchemeGroupVersion = schema.GroupVersion{Group: "rocketmq.apache.org", Version: "v1alpha1"}
 
 	// SchemeBuilder is used to add go types to the GroupVersionKind scheme
 	SchemeBuilder = &scheme.Builder{GroupVersion: SchemeGroupVersion}
